# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/branches": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene todas las sucursales con filtros y paginación",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "branches"
                ],
                "summary": "Obtiene todas las sucursales de la empresa del usuario autenticado",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Número de página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Tamaño de página",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Nombre de la sucursal",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Código de la sucursal",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Nombre del contacto",
                        "name": "contact_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Email del contacto",
                        "name": "contact_email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ID de la zona",
                        "name": "zone_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Estado (activo/inactivo)",
                        "name": "is_active",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Campo por el cual ordenar",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Dirección de ordenamiento (asc/desc)",
                        "name": "sort_direction",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.PaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea una nueva sucursal para la empresa del usuario autenticado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "branches"
                ],
                "summary": "Crea una nueva sucursal",
                "parameters": [
                    {
                        "description": "Información de la sucursal",
                        "name": "branch",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.BranchCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Sucursal creada exitosamente",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/branches/available-zones/{branch_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna las zonas que pueden ser asignadas a una sucursal específica",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "branches",
                    "zones"
                ],
                "summary": "Obtiene las zonas disponibles para una sucursal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la sucursal",
                        "name": "branch_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_domain_delivery_models_entities.Zone"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/branches/deactivate/{branch_id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cambia el estado de una sucursal a inactivo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "branches"
                ],
                "summary": "Desactiva una sucursal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la sucursal",
                        "name": "branch_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucursal desactivada exitosamente",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/branches/reactivate/{branch_id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cambia el estado de una sucursal a activo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "branches"
                ],
                "summary": "Reactiva una sucursal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la sucursal",
                        "name": "branch_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucursal reactivada exitosamente",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/branches/zones/{branch_id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Asigna una zona geográfica específica a una sucursal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "branches",
                    "zones"
                ],
                "summary": "Asigna una zona a una sucursal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la sucursal",
                        "name": "branch_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Información de la zona a asignar",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.ZoneAssignmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Zona asignada exitosamente",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/branches/{branch_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene los detalles de una sucursal específica",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "branches"
                ],
                "summary": "Obtiene una sucursal por su ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la sucursal",
                        "name": "branch_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.BranchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Actualiza la información de una sucursal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "branches"
                ],
                "summary": "Actualiza una sucursal existente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la sucursal",
                        "name": "branch_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Información actualizada de la sucursal",
                        "name": "branch",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.BranchUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucursal actualizada exitosamente",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_MarlonG1_delivery-backend_internal_domain_delivery_models_entities.Zone": {
            "type": "object",
            "properties": {
                "base_rate": {
                    "type": "number"
                },
                "boundaries": {
                    "type": "string"
                },
                "center_point": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "max_delivery_time": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "priority_level": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.BranchCreateRequest": {
            "type": "object",
            "required": [
                "code",
                "contact_email",
                "contact_name",
                "contact_phone",
                "name",
                "operating_hours",
                "zone_id"
            ],
            "properties": {
                "address": {
                    "description": "Dirección asociada a la sucursal (opcional)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.CompanyAddressDTO"
                        }
                    ]
                },
                "code": {
                    "description": "Código único de la sucursal",
                    "type": "string",
                    "example": "SUC-NORTE-001"
                },
                "contact_email": {
                    "description": "Email del contacto en la sucursal",
                    "type": "string",
                    "example": "norte@expressdelivery.com"
                },
                "contact_name": {
                    "description": "Nombre del contacto en la sucursal",
                    "type": "string",
                    "example": "Gerente Norte"
                },
                "contact_phone": {
                    "description": "Teléfono del contacto en la sucursal",
                    "type": "string",
                    "example": "+573001112233"
                },
                "name": {
                    "description": "Nombre de la sucursal",
                    "type": "string",
                    "example": "Sucursal Norte"
                },
                "operating_hours": {
                    "description": "Horarios de operación de la sucursal",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.OperatingHoursDTO"
                        }
                    ]
                },
                "zone_id": {
                    "description": "ID de la zona a la que pertenece la sucursal",
                    "type": "string",
                    "example": "f8c3e8d7-b6a5-4d3c-9f1e-0a2b4c6d8e0f"
                }
            }
        },
...
        "github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.BranchUpdateRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Código único de la sucursal",
                    "type": "string",
                    "example": "SUC-NORTE-001"
                },
                "contact_email": {
                    "description": "Email del contacto en la sucursal",
                    "type": "string",
                    "example": "norte@expressdelivery.com"
                },
                "contact_name": {
                    "description": "Nombre del contacto en la sucursal",
                    "type": "string",
                    "example": "Gerente Norte"
                },
                "contact_phone": {
                    "description": "Teléfono del contacto en la sucursal",
                    "type": "string",
                    "example": "+573001112233"
                },
                "name": {
                    "description": "Nombre de la sucursal",
                    "type": "string",
                    "example": "Sucursal Norte"
                },
                "operating_hours": {
                    "description": "Horarios de operación de la sucursal",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.OperatingHoursDTO"
                        }
                    ]
                },
                "zone_id": {
                    "description": "ID de la zona a la que pertenece la sucursal",
                    "type": "string",
                    "example": "f8c3e8d7-b6a5-4d3c-9f1e-0a2b4c6d8e0f"
                }
            }
        },
...
        "github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.OperatingHoursDTO": {
            "type": "object",
            "required": [
                "weekdays",
                "weekends"
            ],
            "properties": {
                "weekdays": {
                    "description": "Horario para días de semana",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.OperatingHoursTimeDTO"
                        }
                    ]
                },
                "weekends": {
                    "description": "Horario para fines de semana",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.OperatingHoursTimeDTO"
                        }
                    ]
                }
            }
        },
...
        "github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.PaginatedResponse": {
            "description": "Paginated data response with metadata about pagination",
            "type": "object",
            "properties": {
                "data": {
                    "description": "The actual data items (array of objects)"
                },
                "page": {
                    "description": "Current page number",
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "description": "Number of items per page",
                    "type": "integer",
                    "example": 20
                },
                "total_items": {
                    "description": "Total number of items across all pages",
                    "type": "integer",
                    "example": 100
                },
                "total_pages": {
                    "description": "Total number of pages",
                    "type": "integer",
                    "example": 5
                }
            }
        },
...
        "github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.ZoneAssignmentRequest": {
            "description": "Solicitud para asignar una zona a una sucursal",
            "type": "object",
            "required": [
                "zone_id"
            ],
            "properties": {
                "zone_id": {
                    "description": "ID de la zona a asignar\n@required",
                    "type": "string",
                    "example": "f8c3e8d7-b6a5-4d3c-9f1e-0a2b4c6d8e0f"
                }
            }
        },
        "github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIError"
                }
            }
        },
...
    }
}